(PC-41720)[PRO] feat: reorganize user folder and migrate to newApi
