+++ pro/src/pages/User/UserProfile/BannerPendingEmailValidation/BannerPendingEmailValidation.module.scss
