+++ pro/src/pages/User/UserProfile/BannerPendingEmailValidation/BannerPendingEmailValidation.spec.tsx
