+++ pro/src/pages/User/UserProfile/BannerPendingEmailValidation/BannerPendingEmailValidation.tsx
